requirements.txt
# ── Core ML ──────────────────────────────────────────────────────
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.4.0
xgboost==2.0.3
imbalanced-learn==0.11.0
joblib==1.3.2

# ── Experiment Tracking ───────────────────────────────────────────
mlflow==2.9.2

# ── Data Versioning ───────────────────────────────────────────────
dvc==3.38.1
dvc-azure==3.1.0

# ── API Serving ───────────────────────────────────────────────────
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3

# ── Monitoring & Drift ────────────────────────────────────────────
evidently==0.4.13

# ── Utilities ─────────────────────────────────────────────────────
python-dotenv==1.0.0
pyyaml==6.0.1

# ── Testing ───────────────────────────────────────────────────────
pytest==7.4.4
httpx==0.26.0

# ── Notebooks & Visualization ─────────────────────────────────────
jupyter==1.0.0
matplotlib==3.8.2
seaborn==0.13.1
